requirements.txt
Flask==2.1.1
gevent==21.12.0
h5py==3.6.0
numpy==1.21.6
Pillow==9.1.0
tensorflow==2.8.0
Werkzeug==2.1.1
WSGIserver==1.3
PyQt5==5.15.6
PyQtWebEngine==5.15.5
python-dotenv==0.19.2
